refactor: further split use-project-lifecycle into focused hooks (#748)
Extract remaining concerns from use-project-lifecycle.ts into:
- use-project-initialization.ts: loadProjectContent + openRestoredProject (VFS/config loading)
- use-file-opening.ts: handleOpenProject, handleOpenStandaloneFile,
  handleOpenRecentProject, handleOpenAsProject (open dialog logic)

use-project-lifecycle is now a thin orchestration layer (~90 lines of logic)
delegating to six focused sub-hooks.
